requirements.txt
flask
redis
numpy
pillow
pytesseract
opencv-python
unidecode